backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.2
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-environ==0.4.54
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
